Fix initial setup to fetch everything
Initially, all entities will not be part of the entity registry, so
nothing would be fetched. This was overlooked during testing because the
entity registry is not cleared when a config entry is unloaded.
Reconfiguring to test leaves values in the registry, so all will work as
desired when unloading and reloading an existing config entry.

This fixes the initial case, though, by simply inverting the logic in
each of the cases.