Add restoration of `unit_system` state
This was not done before, so if the API returned imperial data, this
would not be converted correctly after a restart.

This also ensures that the data is cleared out when the header is absent
in later requests (which is unlikely).

It also looks up the `unit_system` from the `value_key_path`, ensuring
that the unit system is used for all sensors that the data applies to.